Put Stats button in it's own file instead of putting it in the main file so we could only display it when we run the game
